🏗️ Move components inside `src/components/Buttons/` up one level
This `Buttons` folder was prone to confusion, since this app doesn't use custom UI components but relies on MUI instead.
