+++ src/components/BasicButton/style.scss
@@ -1,4 +1,4 @@
-@use '../../../styles/settings/colors' as v;
+@use '../../styles/settings/colors' as v;
 
 .c-btn {
     color: v.$light-secondary;
