+++ src/components/Nav/ButtonHome/index.jsx
@@ -3,7 +3,7 @@ import { Link, useLocation } from 'react-router-dom'
 import Button from '@mui/material/Button'
 import { Box } from '@mui/material'
 import { HashLink } from 'react-router-hash-link'
-import { BasicButton } from '../../Buttons/BasicButton'
+import { BasicButton } from '../../BasicButton'
 import { selectUserIsLogged, selectUserId, selectUserOrganizationId } from '../../../features/user/store/userSelectors'
 import { selectErrorPageCode } from '../../../store/errorPageSelectors'
 import { ErrorCode } from '../../../store/errorPageSlice'
