+++ src/features/admin/pages/Administration/index.jsx
@@ -1,6 +1,6 @@
 import { AuthenticatedLayout } from '../../../../layouts/AuthenticatedLayout'
 import { MembersList } from '../../members-list/components/MembersList'
-import { ScrollTopButton } from '../../../../components/Buttons/ScrollTopButton'
+import { ScrollTopButton } from '../../../../components/ScrollTopButton'
 import './style.scss'
 import { MembersListProvider } from '../../members-list/contexts/MembersListProvider'
