+++ src/features/user/pages/UserProfile/index.jsx
@@ -1,5 +1,5 @@
 import { ActivityFeed } from '../../../activity-feed/post/components/ActivityFeed'
-import { ScrollTopButton } from '../../../../components/Buttons/ScrollTopButton'
+import { ScrollTopButton } from '../../../../components/ScrollTopButton'
 import { AuthenticatedLayout } from '../../../../layouts/AuthenticatedLayout'
 import { ProfileUserIdProvider } from '../../../activity-feed/post/contexts/ProfileUserIdProvider'
 import './style.scss'
